AP_PitchController: Changes arising from review

+++ libraries/APM_Control/AP_PitchController.h
@@ -31,7 +31,7 @@ public:
     void autotune_start(void) { autotune.start(); }
     void autotune_restore(void) { autotune.stop(); }
 
-    // When called, this function causes the autopilot gain to be increased by the amount set by PTCH2SRV_FGAIN
+    // When called, this function causes the autopilot gain to be increased by the amount set by PTCH2SRV_FGS_MAX
     // with the gain increase being applied over a time constant of PTCH2SRV_TCONST. The Command times out after
     // 100msec so it should be called on every frame of the attitude controller.
     void doing_flare(void) { _flare_active_time_ms = AP_HAL::millis(); }
@@ -72,6 +72,6 @@ private:
     // flare gain management
     bool _flare_active_time_ms = 0;
     AP_Float _flare_gain_scaler_max;        // Maximum gain factor applied to feedback loops during flare
-    float _flare_gain_scaler = 1.0f;
+    uint32_t _flare_gain_scaler = 1.0f;     // gain scaler applied during flare manoeuvre
 
 };
